+++ assets/karma.conf.js
@@ -304,7 +304,7 @@ function getCustomLaunchers() {
             browser: 'safari',
             browser_version: 'latest',
             os: 'OS X',
-            os_version: 'High Sierra'
+            os_version: 'Mojave'
         },
 
         // iOS browsers