+++ assets/src/scripts/components/hydrograph/index.js
@@ -1,27 +1,27 @@
 /**
  * Hydrograph charting module.
  */
-import { select } from 'd3-selection';
+import {select} from 'd3-selection';
 
-import { createStructuredSelector } from 'reselect';
+import {createStructuredSelector} from 'reselect';
 
-import { link } from '../../lib/d3-redux';
+import {link} from '../../lib/d3-redux';
+import {isLoadingTS, hasAnyTimeSeries} from '../../selectors/time-series-selector';
+import {Actions} from '../../store';
 
-import {isLoadingTS, hasAnyTimeSeries } from '../../selectors/time-series-selector';
-import { Actions } from '../../store';
-
-import { cursorSlider } from './cursor';
-import { drawDateRangeControls } from './date-controls';
-import { lineSegmentsByParmCdSelector } from './drawing-data';
-import { drawGraphControls } from './graph-controls';
-import { SPARK_LINE_DIM, layoutSelector } from './layout';
-import { drawTimeSeriesLegend } from './legend';
-import { drawLoadingIndicator } from '../loading-indicator';
-import { drawMethodPicker } from './method-picker';
-import { plotSeriesSelectTable, availableTimeSeriesSelector } from './parameters';
-import { timeSeriesScalesByParmCdSelector } from './scales';
-import { allTimeSeriesSelector } from './time-series';
-import { drawTimeSeriesGraph } from './time-series-graph';
+import {cursorSlider} from './cursor';
+import {drawDateRangeControls} from './date-controls';
+import {lineSegmentsByParmCdSelector} from './drawing-data';
+import {drawGraphBrush} from './graph-brush';
+import {drawGraphControls} from './graph-controls';
+import {SPARK_LINE_DIM}  from './layout';
+import {drawTimeSeriesLegend} from './legend';
+import {drawLoadingIndicator} from '../loading-indicator';
+import {drawMethodPicker} from './method-picker';
+import {plotSeriesSelectTable, availableTimeSeriesSelector} from './parameters';
+import {timeSeriesScalesByParmCdSelector} from './scales';
+import {allTimeSeriesSelector} from './time-series';
+import {drawTimeSeriesGraph} from './time-series-graph';
 
 
 const drawMessage = function(elem, message) {
@@ -104,10 +104,10 @@ export const attachToNode = function (store,
     // Fetch the time series data
     if (period) {
         store.dispatch(Actions.retrieveCustomTimePeriodTimeSeries(siteno, parameter ? parameter : '00060', period))
-            .catch((message) => dataLoadingAlert(select(node), message ? message : 'No data returned'));
+            .catch((message) => dataLoadingAlert(nodeElem, message ? message : 'No data returned'));
     } else {
         store.dispatch(Actions.retrieveTimeSeries(siteno, parameter ? [parameter] : null))
-            .catch(() => dataLoadingAlert((select(node), 'No current time series data available for this site')));
+            .catch(() => dataLoadingAlert((nodeElem, 'No current time series data available for this site')));
     }
     store.dispatch(Actions.retrieveMedianStatistics(siteno));
 
@@ -118,6 +118,7 @@ export const attachToNode = function (store,
     if (!showOnlyGraph) {
         graphContainer.call(cursorSlider, store);
     }
+    graphContainer.call(drawGraphBrush, store);
     graphContainer.append('div')
         .classed('ts-legend-controls-container', true)
         .call(drawTimeSeriesLegend, store);
@@ -136,8 +137,7 @@ export const attachToNode = function (store,
                 siteno: () => siteno,
                 availableTimeSeries: availableTimeSeriesSelector,
                 lineSegmentsByParmCd: lineSegmentsByParmCdSelector('current', 'P7D'),
-                timeSeriesScalesByParmCd: timeSeriesScalesByParmCdSelector('current', 'P7D', SPARK_LINE_DIM),
-                layout: layoutSelector
+                timeSeriesScalesByParmCd: timeSeriesScalesByParmCdSelector('current', 'P7D', SPARK_LINE_DIM)
             }), store));
         nodeElem.select('.provisional-data-alert')
             .call(link(store, function(elem, allTimeSeries) {