+++ assets/src/scripts/components/hydrograph/scales.js
@@ -2,7 +2,7 @@ import { scaleLinear, scaleSymlog } from 'd3-scale';
 import memoize from 'fast-memoize';
 import { createSelector } from 'reselect';
 import { getYDomain, SYMLOG_PARMS } from './domain';
-import { layoutSelector } from './layout';
+import { getLayout } from './layout';
 import { timeSeriesSelector, TEMPERATURE_PARAMETERS } from './time-series';
 import { visiblePointsSelector, pointsByTsKeySelector } from './drawing-data';
 import { getVariables, getCurrentParmCd, getRequestTimeRange } from '../../selectors/time-series-selector';
@@ -64,13 +64,24 @@ export const createYScale = function (parmCd, extent, size) {
  * @param  {Object} state       Redux store
  * @return {Function}           D3 scale function
  */
-export const xScaleSelector = memoize(tsKey => createSelector(
-    layoutSelector,
+export const getXScale = memoize((kind, tsKey) => createSelector(
+    getLayout(kind),
     getRequestTimeRange(tsKey),
-    (layout, requestTimeRange) => {
-        return createXScale(requestTimeRange, layout.width - layout.margin.right);
+        state => state.ui.hydrographXRange,
+    (layout, requestTimeRange, hydrographXRange) => {
+        let timeRange;
+        if (kind === 'BRUSH') {
+            timeRange = requestTimeRange;
+        } else {
+            timeRange = hydrographXRange ? hydrographXRange : requestTimeRange;
+        }
+        return createXScale(timeRange, layout.width - layout.margin.right);
     }
 ));
+
+export const getMainXScale = (tsKey) => getXScale('MAIN', tsKey);
+export const getBrushXScale = (tsKey) => getXScale('BRUSH', tsKey);
+
 
 
 /**
@@ -78,18 +89,21 @@ export const xScaleSelector = memoize(tsKey => createSelector(
  * @param  {Object} state   Redux store
  * @return {Function}       D3 scale function
  */
-export const yScaleSelector = createSelector(
-    layoutSelector,
+export const getYScale = memoize(kind => createSelector(
+    getLayout(kind),
     visiblePointsSelector,
     getCurrentParmCd,
     (layout, pointArrays, currentVarParmCd) => {
         const yDomain = getYDomain(pointArrays, currentVarParmCd);
         return createYScale(currentVarParmCd, yDomain, layout.height - (layout.margin.top + layout.margin.bottom));
     }
-);
+));
 
-export const secondaryYScaleSelector = createSelector(
-    layoutSelector,
+export const getMainYScale = getYScale();
+export const getBrushYScale = getYScale('BRUSH');
+
+export const getSecondaryYScale = memoize(kind => createSelector(
+    getLayout(kind),
     visiblePointsSelector,
     getCurrentParmCd,
     (layout, pointArrays, currentVarParmCd) => {
@@ -106,7 +120,7 @@ export const secondaryYScaleSelector = createSelector(
             currentVarParmCd, convertedYDomain, layout.height - (layout.margin.top + layout.margin.bottom)
         );
     }
-);
+));
 
 
 /**