+++ assets/src/scripts/store/index.js
@@ -369,6 +369,17 @@ export const Actions = {
             width
         };
     },
+    setHydrographXRange(hydrographXRange) {
+        return {
+            type: 'SET_HYDROGRAPH_X_RANGE',
+            hydrographXRange
+        };
+    },
+    clearHydrographXRange() {
+        return {
+            type: 'CLEAR_HYDROGRAPH_X_RANGE'
+        };
+    },
     setCurrentVariable(variableID) {
         return {
             type: 'SET_CURRENT_VARIABLE',
@@ -400,7 +411,7 @@ export const Actions = {
             const locationIanaTimeZone = getIanaTimeZone(state);
             const startTime = new DateTime.fromISO(startTimeStr,{zone: locationIanaTimeZone}).toMillis();
             const endTime = new DateTime.fromISO(endTimeStr, {zone: locationIanaTimeZone}).toMillis();
-            dispatch(Actions.retrieveCustomTimeSeries(siteno, startTime, endTime));
+            return dispatch(Actions.retrieveCustomTimeSeries(siteno, startTime, endTime));
         };
     },
     setGageHeightFromStageIndex(index) {
@@ -472,7 +483,8 @@ export const configureStore = function (initialState) {
         },
         ui : {
             windowWidth: 1024,
-            width: 800
+            width: 800,
+            hydrographXRange: {}
         },
         ...initialState
     };